refactor: update MongoDB package references and improve test error messages
