+++ mongo/mongo.go
@@ -1,4 +1,4 @@
-package database
+package mongo
 
 import (
 	"context"
